feat(users): refactor user authentication and management by replacing AuthUsers module with Users module
